pyproject.toml
[tool.poetry]
name = "pyimg3lib"
version = "1.1.6"
description = "iOS Image3 lib"
authors = ["Merculous <24739823+Merculous@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "img3lib"}]
repository = "https://github.com/Merculous/PyImg3lib"

[tool.poetry.dependencies]
python = "^3.11"
pylzss = "^0.3.4"
pycryptodome = "^3.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
